+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.3.0
+
+- Add setCharacteristicSubscriptionChangeCallback support on Android
+- Breaking Change: Removed setBleCentralAvailabilityCallback ( because all platforms supports characteristics subscription change now )
+- Add setInstance for testing or Mock implementation
+- Add BlePeripheralInterface for setting your own implementation
+- Remove Linux dependency to fix crash on linux
+
 ## 2.2.3
 
 - Fix windows crash on stopping advertisement sometimes